minor bug fix in train

+++ projects/test/train/metrics.py
@@ -131,7 +131,7 @@ class OnlinePsdRatio(Metric):
 
         # postprocess, doing the bandpass filtering back
         # in numpy because torchaudio won't work
-        noise = torch.stack(noise).double()
+        noise = torch.stack(noise)
         noise = self.y_scaler(noise, reverse=True)
         noise = self.bandpass(noise.cpu().detach().numpy())
         noise = torch.tensor(noise, device=device)
@@ -145,7 +145,7 @@ class OnlinePsdRatio(Metric):
         # reshape our raw strain into a timeseries
         raw = torch.cat(self.strain, dim=0)[1 : num_frames - 1]
         raw = raw.view(1, -1)
-        return noise, raw.double()
+        return noise, raw
 
     def compute(self, reduce: bool = True):
         noise, raw = self.clean()
